Rewrite Switch
